Implemented tree traversals, I sure hope this works

+++ Phase2Main.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main(){
 	string K[10] = {"A","B","C","D","E","F","G","H","I","K","L","M"};
-	int V[10] = {12,11,10,9,8,7,6,5,4,3,2,1};
+	float V[10] = {12,11,10,9,8,7,6,5,4,3,2,1};
 	
 	two4Tree<string, float> T1(K,V,12), T2;
